refactor: 보드 메서드 분리 및 위치 변경
